requirements.txt
numpy
Cython
h5py==2.10
Pillow
six
scipy
opencv-python
matplotlib
tb-nightly
future
yacs
gdown
flake8
yapf
isort
imageio
tensorflow==1.15
tensorflow-gpu==1.15
scikit-learn==0.22.2
torch
torchvision
# keras==2.4.0
keras==2.3.1